Add files via upload
Maze game with hidden monsters:
Users are required to choose the direction they want to go in order to reach the end of maze. However, there are hidden monsters along the way, and when the player enters a "monster occupied zone", they are required to attack the monster or flee.